TestFX suite added
- TestFX for ownNote layout added
- Project re-structured to fit standard gradle
- Various minor fixes / updates as result from testing
- Changed tab colors to be similar to OneNote
