+++ ownNoteEditor/build.gradle
@@ -14,7 +14,7 @@ plugins {
 
 macAppBundle {
     mainClassName = "tf.ownnote.ui.main.OwnNoteEditorManager"
-    icon          = "src/tf/ownnote/ui/main/ownNoteEditor.icns"
+    icon          = "src/main/resources/ownNoteEditor.icns"
     volumeName    = "ownNoteEditor"
     dmgName       = "ownNoteEditor"
     javaProperties.put("apple.laf.useScreenMenuBar", "true")
@@ -27,19 +27,6 @@ apply plugin: 'java'
 apply plugin: 'application'
 
 mainClassName = 'tf.ownnote.ui.main.OwnNoteEditorManager'
-
-//adjust location of src and resources
-// TODO this should be restructured to the standard structure like /src/main/java and ../resources later
-sourceSets {
-    main {
-        java {
-            srcDir 'src/'
-        }
-        resources {
-            srcDir 'src/'
-        }
-    }
-}
 
 // In this section you declare where to find the dependencies of your project
 repositories {
@@ -56,6 +43,8 @@ dependencies {
     compile 'commons-io:commons-io:2.4'
     compile 'org.jsoup:jsoup:1.8.3'
 
+    testCompile "org.testfx:testfx-junit:4.0.+"
+    testCompile "org.testfx:testfx-core:4.0.+"
     // Declare the dependency for your favourite test framework you want to use in your tests.
     // TestNG is also supported by the Gradle Test task. Just change the
     // testCompile dependency to testCompile 'org.testng:testng:6.8.1' and add
