+++ ownNoteEditor/src/main/java/tf/ownnote/ui/helper/FormatHelper.java
