+++ ownNoteEditor/src/main/java/tf/ownnote/ui/helper/NoteData.java
