+++ ownNoteEditor/src/main/java/tf/ownnote/ui/helper/OwnNoteDirectoryMonitor.java
