+++ ownNoteEditor/src/main/java/tf/ownnote/ui/helper/OwnNoteHTMLEditor.java
@@ -25,8 +25,11 @@
  */
 package tf.ownnote.ui.helper;
 
+import com.sun.javafx.webkit.Accessor;
+import com.sun.webkit.WebPage;
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
@@ -38,11 +41,20 @@ import javafx.application.Platform;
 import javafx.beans.property.BooleanProperty;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
+import javafx.event.Event;
+import javafx.event.EventHandler;
+import javafx.geometry.Insets;
 import javafx.print.PrinterJob;
+import javafx.scene.Group;
 import javafx.scene.Node;
+import javafx.scene.Scene;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonBase;
 import javafx.scene.control.Separator;
+import javafx.scene.control.TextArea;
+import javafx.scene.control.ToggleButton;
 import javafx.scene.control.ToolBar;
+import javafx.scene.control.Tooltip;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.KeyCode;
@@ -50,9 +62,13 @@ import javafx.scene.input.KeyCodeCombination;
 import javafx.scene.input.KeyCombination;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Priority;
+import javafx.scene.layout.StackPane;
 import javafx.scene.web.HTMLEditor;
+import javafx.scene.web.WebEngine;
 import javafx.scene.web.WebView;
 import javafx.stage.FileChooser;
+import javafx.stage.Stage;
+import javafx.stage.WindowEvent;
 import org.apache.commons.io.FilenameUtils;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
@@ -66,13 +82,26 @@ public class OwnNoteHTMLEditor {
     private static final String TOP_TOOLBAR = ".top-toolbar";
     private static final String BOTTOM_TOOLBAR = ".bottom-toolbar";
     private static final String WEB_VIEW = ".web-view";
+    
+    private static enum ButtonType {
+        ClickButton,
+        ToggleButton
+    };
+    
+    private static final String UNDO_COMMAND = "undo";
+    private static final String REDO_COMMAND = "redo";
+    ButtonBase undoEdit = null;
+    ButtonBase redoEdit = null;
 
     private WebView mWebView;
+    private WebEngine mWebEngine;
+    private WebPage mWebPage;
     private ToolBar mTopToolBar;
     private ToolBar mBottomToolBar;    
     
-    
     private HTMLEditor myHTMLEditor = null;
+    
+    private RawViewer rawViewer =null;
     
     // callback to OwnNoteEditor required for e.g. delete & rename
     private OwnNoteEditor myEditor= null;
@@ -87,6 +116,12 @@ public class OwnNoteHTMLEditor {
         super();
         myHTMLEditor = htmlEditor;
         myEditor = editor;
+
+        mWebView = (WebView) myHTMLEditor.lookup(WEB_VIEW);
+        mWebEngine = mWebView.getEngine();
+        mWebPage = Accessor.getPageFor(mWebEngine);
+        
+        rawViewer = new RawViewer();
         
         // delay setup of editor - things are not available at startup...
         Platform.runLater(() -> {
@@ -95,16 +130,34 @@ public class OwnNoteHTMLEditor {
     }
     
     public void setNoteText(final String text) {
-        myHTMLEditor.setHtmlText(text);
+        myHTMLEditor.setHtmlText(wrapText(text));
         noteText = getNoteText();
+        
+        setUndoRedo();
+    }
+    
+    private String wrapText(final String text) {
+        // chance to add, whatever we would like to have...
+        return String.format(
+                "<!DOCTYPE html>\r\n"
+                + "<html>\r\n"
+                + "  <head>\r\n"
+                + "    <script>\r\n"
+                + "    </script>\r\n"
+                + "  </head>\r\n"
+                + "  <body>%s</body>\r\n"
+                + "</html>", text);
     }
 
     public String getNoteText() {
         String noteHtml = myHTMLEditor.getHtmlText();
         
         Document doc = Jsoup.parse(noteHtml);
+        // Issue #55: seems we now also get another font tag that we don't want to have: "<span style="font-family: 'Segoe UI';">"
+        doc.getElementsByAttributeValueStarting("style", "font-family").unwrap();
         // get rid of "<font face="Segoe UI">" tags - see bug report https://bugs.openjdk.java.net/browse/JDK-8133833
-        doc.getElementsByTag("font").unwrap();
+        // TF, 20160724: match only font + face since we also want to allow foreground / background colors
+        doc.getElementsByAttributeValue("face", "Segoe UI").unwrap();
         // only store content in <body>
         noteHtml = doc.select("body").html();
         
@@ -168,19 +221,20 @@ public class OwnNoteHTMLEditor {
     }
 
     private void initHTMLEditor() {
-        mWebView = (WebView) myHTMLEditor.lookup(WEB_VIEW);
         mTopToolBar = (ToolBar) myHTMLEditor.lookup(TOP_TOOLBAR);
         mBottomToolBar = (ToolBar) myHTMLEditor.lookup(BOTTOM_TOOLBAR);
         
+        // TF, 20160724: why should we remove that editor feature?
         // remove: foreground & background control
-        hideNode(myHTMLEditor, ".html-editor-foreground", 1);
-        hideNode(myHTMLEditor, ".html-editor-background", 1);
+        // hideNode(myHTMLEditor, ".html-editor-foreground", 1);
+        // hideNode(myHTMLEditor, ".html-editor-background", 1);
+        
         // remove: font type & font size control - the 2nd and 3rd control with "font-menu-button" style class
         hideNode(myHTMLEditor, ".font-menu-button", 2);
         hideNode(myHTMLEditor, ".font-menu-button", 3);
+        
         // add: insert link & picture controls
         addNoteEditorControls();
-        // add: undo & redo button, back button
         
         // issue #40
         // https://stackoverflow.com/questions/20773249/javafx-htmleditor-doesnt-take-all-free-size-on-the-container
@@ -220,7 +274,7 @@ public class OwnNoteHTMLEditor {
         //     event.consume();
         // });
     }
-    
+
     private static void hideNode(final Node startNode, final String lookupString, final int occurence) {
         final Set<Node> nodes = startNode.lookupAll(lookupString);
         if (nodes != null && nodes.size() >= occurence) {
@@ -244,22 +298,27 @@ public class OwnNoteHTMLEditor {
         }
 
         // copy styles from other buttons in toolbar
-        ObservableList<String> buttonStyles = null;
+        ObservableList<String> clickButtonStyles = null;
         Node node = mTopToolBar.lookup(".html-editor-cut");
         if (node != null && node instanceof Button) {
-            buttonStyles = ((Button) node).getStyleClass();
+            clickButtonStyles = ((Button) node).getStyleClass();
             // not the own button style, please
-            buttonStyles.removeAll("html-editor-cut");
+            clickButtonStyles.removeAll("html-editor-cut");
+        }
+        ObservableList<String> toggleButtonStyles = null;
+        node = mTopToolBar.lookup(".html-editor-bold");
+        if (node != null && node instanceof Button) {
+            toggleButtonStyles = ((Button) node).getStyleClass();
+            // not the own button style, please
+            toggleButtonStyles.removeAll("html-editor-bold");
         }
 
+        // add separator
+        mBottomToolBar.getItems().add(new Separator());
+        
         // add button to insert link
-        ImageView graphic =
-                new ImageView(new Image(OwnNoteEditor.class.getResourceAsStream("/tf/ownnote/ui/css/link.png"), 22, 22, true, true));
-        final Button insertLink = new Button("", graphic);
-        insertLink.getStyleClass().add("html-editor-insertlink");
-        if (buttonStyles != null) {
-            insertLink.getStyleClass().addAll(buttonStyles);
-        }
+        final ButtonBase insertLink = 
+                    createEditorButton(ButtonType.ClickButton, "link.png", "Insert Link", "html-editor-insertlink", clickButtonStyles);
         insertLink.setOnAction((ActionEvent t) -> {
             LinkDialog linkDialog = new LinkDialog();
             if (linkDialog.showAndWait()) {
@@ -275,20 +334,16 @@ public class OwnNoteHTMLEditor {
                             linkDialog.getWindowMode()+
                             "\">" +
                             linkDialog.getLinkText().trim() + "</a>";
-                    myHTMLEditor.setHtmlText(myHTMLEditor.getHtmlText() + hrefString);
+                    insertHtmlAfterCursor(hrefString);
                 }
             }
+            myHTMLEditor.requestFocus();
         });
-        mTopToolBar.getItems().add(insertLink);
+        mBottomToolBar.getItems().add(insertLink);
 
         // add button to insert image
-        graphic = 
-                new ImageView(new Image(OwnNoteEditor.class.getResourceAsStream("/tf/ownnote/ui/css/insertimage.gif"), 22, 22, true, true));
-        final Button insertImage = new Button("", graphic);
-        insertImage.getStyleClass().add("html-editor-insertimage");
-        if (buttonStyles != null) {
-            insertImage.getStyleClass().addAll(buttonStyles);
-        }
+        final ButtonBase insertImage = 
+                createEditorButton(ButtonType.ClickButton, "insertimage.gif", "Insert Image", "html-editor-insertimage", clickButtonStyles);
         insertImage.setOnAction((ActionEvent arg0) -> {
             final List<String> extFilter = Arrays.asList("*.jpg", "*.png", "*.gif");
             final List<String> extValues = Arrays.asList("jpg", "png", "gif");
@@ -307,49 +362,169 @@ public class OwnNoteHTMLEditor {
                     importDataFile(selectedFile);
                 }                        
             }
+            myHTMLEditor.requestFocus();
         });
-        mTopToolBar.getItems().add(insertImage);
+        mBottomToolBar.getItems().add(insertImage);
 
         // add separator
-        mTopToolBar.getItems().add(new Separator());
+        mBottomToolBar.getItems().add(new Separator());
+
+        // add source button
+        final ToggleButton viewSource = 
+                (ToggleButton) createEditorButton(ButtonType.ToggleButton, "source.png", "View Source", "html-editor-bold", toggleButtonStyles);
+        viewSource.setOnAction((ActionEvent arg0) -> {
+            if (viewSource.isSelected()) {
+                rawViewer.show();
+                rawViewer.setText(getNoteText());
+            } else {
+                rawViewer.hide();
+            }
+        });
+        // uncheck button when window gets closed
+        rawViewer.setOnCloseRequest(new EventHandler<WindowEvent>() {
+            @Override
+            public void handle(WindowEvent event) {
+                viewSource.setSelected(false);
+            }
+        });        
+        mBottomToolBar.getItems().add(viewSource);
+
+        // add separator
+        mBottomToolBar.getItems().add(new Separator());
 
         // Issue #12 - add print button
         // https://stackoverflow.com/questions/28847757/how-to-display-print-dialog-in-java-fx-and-print-node
-        graphic =
-                new ImageView(new Image(OwnNoteEditor.class.getResourceAsStream("/tf/ownnote/ui/css/print.png"), 22, 22, true, true));
-        final Button printNote = new Button("", graphic);
-        printNote.getStyleClass().add("html-editor-print");
-        if (buttonStyles != null) {
-            printNote.getStyleClass().addAll(buttonStyles);
-        }
+        final ButtonBase printNote = 
+                createEditorButton(ButtonType.ClickButton, "print.png", "Print Note", "html-editor-print", clickButtonStyles);
         printNote.setOnAction((ActionEvent arg0) -> {
             PrinterJob job = PrinterJob.createPrinterJob();
             if (job != null && job.showPrintDialog(myHTMLEditor.getScene().getWindow())){
                 myHTMLEditor.print(job);
                 job.endJob();
             }
+            myHTMLEditor.requestFocus();
         });
-        mTopToolBar.getItems().add(printNote);
+        mBottomToolBar.getItems().add(printNote);
 
         // add separator
-        mTopToolBar.getItems().add(new Separator());
+        mBottomToolBar.getItems().add(new Separator());
 
         // add save button
-        graphic =
-                new ImageView(new Image(OwnNoteEditor.class.getResourceAsStream("/tf/ownnote/ui/css/save.png"), 22, 22, true, true));
-        final Button saveNote = new Button("", graphic);
-        saveNote.getStyleClass().add("html-editor-print");
-        if (buttonStyles != null) {
-            saveNote.getStyleClass().addAll(buttonStyles);
-        }
+        final ButtonBase saveNote = 
+                createEditorButton(ButtonType.ClickButton, "save.png", "Save Note", "html-editor-save", clickButtonStyles);
         saveNote.setOnAction((ActionEvent arg0) -> {
             saveNote();
+            setUndoRedo();
+            myHTMLEditor.requestFocus();
         });
-        mTopToolBar.getItems().add(saveNote);
+        mBottomToolBar.getItems().add(saveNote);
         // issue #41 - add CTRL+S to button and not to context menu
         saveNote.getScene().getAccelerators().put(new KeyCodeCombination(KeyCode.S, KeyCombination.CONTROL_DOWN), () -> {
             saveNote.fire();
         });
+
+        // issue #56: add undo & redo button
+        // they should go at the beginning of toolbar #1 - so we need to remove everything first...
+        List<Node> toolbarItems = new ArrayList<Node>(mTopToolBar.getItems());
+        mTopToolBar.getItems().clear();
+
+        // add undo button
+        undoEdit = createEditorButton(ButtonType.ClickButton, "undo.png", "Undo", "html-editor-undo", clickButtonStyles);
+        undoEdit.setOnAction((ActionEvent arg0) -> {
+            if (mWebPage.queryCommandEnabled(UNDO_COMMAND)) {
+                final String curText = getNoteText();
+                mWebPage.executeCommand(UNDO_COMMAND, null);
+                // in some cases the first call do UNDO doesn't change anything, only the second... maybe one day I'll find a fix
+                if (curText.equals(getNoteText())) {
+                    mWebPage.executeCommand(UNDO_COMMAND, null);
+                }
+            }
+            setUndoRedo();
+            myHTMLEditor.requestFocus();
+        });
+        mTopToolBar.getItems().add(undoEdit);
+        undoEdit.getScene().getAccelerators().put(new KeyCodeCombination(KeyCode.Z, KeyCombination.CONTROL_DOWN), () -> {
+            undoEdit.fire();
+        });
+        
+        // add redo button
+        redoEdit = createEditorButton(ButtonType.ClickButton, "redo.png", "Redo", "html-editor-redo", clickButtonStyles);
+        redoEdit.setOnAction((ActionEvent arg0) -> {
+            if (mWebPage.queryCommandEnabled(REDO_COMMAND)) {
+                final String curText = getNoteText();
+                mWebPage.executeCommand(REDO_COMMAND, null);
+                // in some cases the first call do UNDO doesn't change anything, only the second... maybe one day I'll find a fix
+                if (curText.equals(getNoteText())) {
+                    mWebPage.executeCommand(REDO_COMMAND, null);
+                }
+            }
+            setUndoRedo();
+            myHTMLEditor.requestFocus();
+        });
+        mTopToolBar.getItems().add(redoEdit);
+        redoEdit.getScene().getAccelerators().put(new KeyCodeCombination(KeyCode.Y, KeyCombination.CONTROL_DOWN), () -> {
+            redoEdit.fire();
+        });
+
+        // add separator
+        mTopToolBar.getItems().add(new Separator());
+
+        // and now once again all other controls
+        mTopToolBar.getItems().addAll(toolbarItems);
+        
+        // TODO: this only catches keys typed but not formatting via buttons - no good idea how this can be monitored
+        // set undo / redo based on command state
+        myHTMLEditor.addEventHandler(Event.ANY, new EventHandler<Event>() {
+            @Override
+            public void handle(Event event) {
+                setUndoRedo();
+                
+                // update raw viewer if visible
+                if (rawViewer.isShowing()) {
+                    rawViewer.setText(getNoteText());
+                }
+            }
+        });
+    }
+    
+    private void setUndoRedo() {
+        if (undoEdit != null && redoEdit != null) {
+            undoEdit.setDisable(!mWebPage.queryCommandEnabled(UNDO_COMMAND));
+            redoEdit.setDisable(!mWebPage.queryCommandEnabled(REDO_COMMAND));
+        }
+    }
+    
+    private ButtonBase createEditorButton(
+            final ButtonType buttonType, 
+            final String iconName, 
+            final String tooltipText, 
+            final String cssStyleClass, 
+            final ObservableList<String> buttonStyles) {
+        final ImageView graphic =
+                new ImageView(new Image(OwnNoteEditor.class.getResourceAsStream("/" + iconName), 20, 20, true, true));
+
+        ButtonBase newButton;
+        switch (buttonType) {
+            case ClickButton: 
+                newButton = new Button("", graphic);
+                break;
+
+            case ToggleButton: 
+                newButton = new ToggleButton("", graphic);
+                break;
+
+            default: 
+                newButton = new Button("", graphic);
+                break;
+        }
+
+        newButton.setTooltip(new Tooltip(tooltipText));
+        newButton.getStyleClass().add(cssStyleClass);
+        if (buttonStyles != null) {
+            newButton.getStyleClass().addAll(buttonStyles);
+        }
+        
+        return newButton;
     }
 
     @SuppressWarnings("unchecked")
@@ -412,7 +587,7 @@ public class OwnNoteHTMLEditor {
                 + "  }"
                 + "})(\"%s\");", html);
         //execute script
-        mWebView.getEngine().executeScript(script);
+        mWebEngine.executeScript(script);
     }
     
     /**
@@ -461,5 +636,49 @@ public class OwnNoteHTMLEditor {
     
     public BooleanProperty visibleProperty() {
         return myHTMLEditor.visibleProperty();
+    }
+    
+    // internal class for raw viewer
+    private class RawViewer {
+        private final Stage stage;
+        private final TextArea htmlCode;
+        
+        public RawViewer() {
+            https://docs.oracle.com/javafx/2/ui_controls/editor.htm
+            stage = new Stage();
+
+            htmlCode = new TextArea();
+            htmlCode.setWrapText(true);
+            htmlCode.setEditable(false);
+
+            final StackPane root = new StackPane();
+            root.setPadding(new Insets(1, 1, 1, 1));
+            root.getChildren().add(htmlCode);
+
+            final Scene scene = new Scene(new Group(), 400, 400);
+            scene.setRoot(root);
+
+            stage.setScene(scene);
+        }
+        
+        public void setText(final String text) {
+            htmlCode.setText(text);
+        }
+        
+        public void show() {
+            stage.show();
+        }
+        
+        public void hide() {
+            stage.hide();
+        }
+        
+        public boolean isShowing() {
+            return stage.isShowing();
+        }
+        
+        public void setOnCloseRequest(final EventHandler<WindowEvent> eventHandler) {
+            stage.setOnCloseRequest(eventHandler);        
+        }
     }
 }
