+++ ownNoteEditor/src/main/java/tf/ownnote/ui/helper/OwnNoteTableColumn.java
