+++ ownNoteEditor/src/main/resources/OwnNoteEditorManager.ico
