+++ ownNoteEditor/src/main/resources/OwnNoteEditorManager.png
