+++ ownNoteEditor/src/main/resources/delete-hover.png
