+++ ownNoteEditor/src/main/resources/delete.png
