+++ ownNoteEditor/src/main/resources/insertimage.gif
