+++ ownNoteEditor/src/main/resources/link.png
