+++ ownNoteEditor/src/main/resources/print.png
