+++ ownNoteEditor/src/main/resources/save.png
