--- ownNoteEditor/src/tf/ownnote/ui/css/ownnote.css
@@ -1,157 +1,0 @@
-/* https://stackoverflow.com/questions/12746459/how-to-change-the-colors-of-the-tableviews-rows */
-/* https://stackoverflow.com/questions/21707178/javafx-how-to-set-selected-row-text-color-in-not-focused-tableview */
-
-/* no focus rectangle and white as default background */
-.root {
-    -fx-background-color: white;
-    -fx-focus-color: transparent;
-    -fx-faint-focus-color: transparent;
-    -fx-border-style: none;
-    -fx-border-width: 0;
-    -fx-font-size: 1.166667em;
-    -fx-font-color: -fx-text-base-color;
-    -fx-text-fill: -fx-text-base-color;
-    
-    /* our own variables should be initialized here as well to avoid run-time warnings */
-    tab-color: white;
-    note-selected-background-color: -fx-base;
-    note-selected-font-color: black;
-    notename-font-weight: bold;
-}
-
-/* No background color in table header, please */
-.table-view.notesTable,
-.table-view.notesTable .column-header,
-.table-view.notesTable .column-header-background,
-.table-view.notesTable .table-row-cell {
-    -fx-background-color: tab-color;
-    -fx-table-cell-border-color: transparent;
-}
-.table-view.notesTable .table-row-cell:hover:filled {
-    -fx-background-color: -fx-base;
-}
-.table-view.notesTable .table-row-cell:selected {
-    -fx-background-color: white;
-}
-.table-view.notesTable .table-row-cell:selected .text {
-    -fx-fill: note-selected-font-color;
-    -fx-font-weight: bold;
-}
-/* note name in bold */
-.noteNameCol {
-    -fx-font-weight: notename-font-weight;
-    -fx-alignment: CENTER_LEFT;
-}
-/* note date in gray */
-.noteModifiedCol {
-    -fx-alignment: CENTER;
-}
-.table-view.notesTable .table-row-cell .noteModifiedCol {
-    -fx-opacity: 0.75;
-}
-/* delete icon only when selected, red on hover */
-.noteDeleteCol:filled {
-    -fx-background-image: url("delete.png");
-    -fx-background-repeat: no-repeat;
-    -fx-background-position: center;
-    -fx-opacity: 0.3;
-}
-.noteDeleteCol:filled:hover {
-    -fx-background-image: url("delete-hover.png");
-    -fx-background-repeat: no-repeat;
-    -fx-background-position: center;
-    -fx-opacity: 1;
-}
-
-/* No background color in table header, please */
-.table-view.groupsTable,
-.table-view.groupsTable .column-header,
-.table-view.groupsTable .column-header-background,
-.table-view.groupsTable .table-row-cell {
-    -fx-background-color: -fx-base;
-    -fx-table-cell-border-color: transparent;
-}
-.table-view.groupsTable .table-row-cell:hover:filled,
-.table-view.groupsTable .table-row-cell:selected {
-    -fx-background-color: darkgray;
-    -fx-font-color: -fx-text-base-color;
-}
-/* selected group info in bold */
-.table-view.groupsTable .table-row-cell:selected {
-    -fx-font-weight: bold;
-}
-.table-view.groupsTable .table-row-cell:selected * {
-    -fx-font-color: -fx-text-base-color;
-    -fx-fill: -fx-text-base-color;
-}
-/* group name left */
-.groupNameCol {
-    -fx-alignment: CENTER_LEFT;
-}
-/* delete icon only when cell has value + is selected, red on hover */
-.table-view.groupsTable .table-row-cell:selected .groupDeleteCol.valueSet:filled {
-    -fx-background-image: url("delete.png");
-    -fx-background-repeat: no-repeat;
-    -fx-background-position: center;
-    -fx-opacity: 0.3;
-}
-.table-view.groupsTable .table-row-cell:selected .groupDeleteCol.valueSet:filled:hover {
-    -fx-background-image: url("delete-hover.png");
-    -fx-background-repeat: no-repeat;
-    -fx-background-position: center;
-    -fx-opacity: 1;
-}
-/* group count left */
-.groupCountCol {
-    -fx-alignment: CENTER_RIGHT;
-}
-
-/* rounded corners on tabs */
-.tab-header-area .tab {
-    -fx-padding: 5 10 5 10;
-    -fx-background-radius: 10 10 0 0;
-}
-.tab:selected .tab-label {
-    -fx-font-weight: bold;
-}
-.tab-pane
-{
-    -fx-tab-min-width: 60px;
-    -fx-border-style: none;
-}
-.tab-pane *.tab-header-background {
-    -fx-background-color: white;
-}
-.tab-content-area 
-{
-    -fx-padding: 0;
-}
-
-/* colors on tabs during run time */
-.tab-header-area .tab {
-    -fx-background-color: tab-color;
-}
-.tab-pane .tab-content-area {
-    -fx-background-color: tab-color;
-}
-
-/* "+" tab is special */
-.tab-header-area .plusTab {
-    -fx-background-color: white;
-    -fx-padding: 4 10 4 10;
-    -fx-border-radius: 10 10 0 0;
-    -fx-border-style: solid;
-    -fx-border-color: darkgray darkgray white darkgray;
-    -fx-font-size: 1.2em;
-    -fx-font-color: darkgray;
-}
-
-/* style the invisible splitpane */
-.split-pane {  
-    -fx-box-border: transparent;  
-    -fx-background-color: transparent;
-}
-.split-pane > .split-pane-divider {  
-    -fx-padding: 0 1 0 1;  
-    -fx-background-color: darkgray;
-}
